% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{abrr_full}
\alias{abrr_full}
\docType{data}
\title{TCGA cancer type abbreviations and full names}
\usage{data(abrr_full)}
\format{A data.frame with 33 rows and 2 columns (tissue, full.name).}
\description{
Mapping between TCGA cancer-type abbreviations (e.g. COAD) and their full names, used when reporting pan-tissue correlation results.
}
\source{TCGA.}
\keyword{datasets}
