% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{ccle_info}
\alias{ccle_info}
\docType{data}
\title{CCLE cell line annotation}
\usage{data(ccle_info)}
\format{A data.frame with 1046 rows and 14 columns.}
\description{
Annotation of the Cancer Cell Line Encyclopedia (CCLE) cell lines, used to map expression samples to cell-line primary sites.
}
\source{\url{https://sites.broadinstitute.org/ccle/}}
\keyword{datasets}
