% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{data_info}
\alias{data_info}
\docType{data}
\title{Description of the TF-target datasets}
\usage{data(data_info)}
\format{A list of 8 entries, each with link and description.}
\description{
Links and descriptions of the databases integrated in the package (hTFtarget, KnockTF, ENCODE, CHEA, TRRUST, GTRD, ChIP_Atlas and JASPAR).
}
\source{\url{https://github.com/WangJin93/TFTF}}
\keyword{datasets}
