% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Find_TF.R
\name{Find_TF}
\alias{Find_TF}
\title{Detect which datasets contain input TFs}
\usage{
Find_TF(tf = "STAT3")
}
\arguments{
\item{tf}{Transcription Factor names.}
}
\description{
Detect which datasets contain input TFs
}
\examples{
\dontrun{
Find_TF(c("STAT3","AHR"))
}
}
