% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_info.R
\name{get_data_info}
\alias{get_data_info}
\title{Get dataset information}
\usage{
get_data_info(dataset = "ENCODE")
}
\arguments{
\item{dataset}{Transcription Factor datasets used in this package, e.g. "ENCODE".}
}
\description{
Get the data information of specific Transcription Factor dataset.
}
\examples{
\dontrun{
get_data_info("ENCODE")
}
}
