% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{get_tftf_cache_dir}
\alias{get_tftf_cache_dir}
\title{Get the TFTF data cache directory}
\usage{
get_tftf_cache_dir()
}
\description{
Print / return the directory used by TFTF to store locally cached downloads (mirroring the cache layout of the GCAS package). The cache directory is resolved in the following order:
\enumerate{
  \item the \code{cache_dir} argument passed to the fetching function;
  \item \code{options(TFTF.cache_dir = ...)};
  \item the \code{TFTF_CACHE_DIR} environment variable;
  \item \code{rappdirs::user_cache_dir("TFTF")} (default).
}
}
\value{
Path to the cache folder (invisibly).
}
\examples{
\dontrun{
get_tftf_cache_dir()
}
}
