% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{idmap}
\alias{idmap}
\docType{data}
\title{Gene symbol to RefSeq/Ensembl identifier mapping}
\usage{data(idmap)}
\format{A data.frame with 61852 rows and 2 columns (ID, gene).}
\description{
Identifier map used to translate between gene symbols and sequence identifiers when querying external TF databases.
}
\source{Compiled for the package.}
\keyword{datasets}
