% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersections.R
\name{intersections}
\alias{intersections}
\title{Get intersections of prediction result}
\usage{
intersections(results, datasets_index = NULL)
}
\arguments{
\item{results}{The result of predict_target() and predict_TF().}

\item{datasets_index}{datasets_index of user inputted datasets. Default NULL, return the intersection of all datasets.}
}
\description{
Get the intersections of the prediction result of predict_target() and predict_TF().
}
\examples{
\dontrun{
results <- predict_target(datasets=c("hTFtarget","KnockTF","FIMO_JASPAR","PWMEnrich_JASPAR"), cor_DB = c("TCGA","GTEx"),tf = "STAT3")
results_inter <- intersections(results)
}
}
