% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{knocktf_data}
\alias{knocktf_data}
\docType{data}
\title{KnockTF experimental annotation}
\usage{data(knocktf_data)}
\format{A data.frame with 1086 rows and 8 columns.}
\description{
Annotation of the KnockTF knockdown/knockout experiments (sample id, TF, molecular type, knock method, biosample, platform, PubMed id ...), merged with KnockTF prediction results in \code{predict_target()} and \code{predict_TF()}.
}
\source{\url{https://bio.liclab.net/KnockTF/}}
\keyword{datasets}
