% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_venn.R
\name{plot_venn}
\alias{plot_venn}
\title{Plot venn diagram and folower plot}
\usage{
plot_venn(
  inter_data,
  color_panel = "Set1",
  font_size = 1,
  lwd = 2,
  linetype = 2
)
}
\arguments{
\item{inter_data}{Intersection result obtained fron intersections() function.}

\item{color_panel}{Color palette in RColorBrewer, default "Set1".}

\item{font_size}{Default 1.}

\item{lwd}{Line width, default 2.}

\item{linetype}{"solid" = 1,"dashed" = 2, "dotted" = 3,"dot-dashed" = 4, "longdashed" = 5, "blank" = 0}
}
\description{
Plot venn diagram and folower plot to visualize the intersection of predicted results.
}
\examples{
\dontrun{
results <- predict_target(datasets=c("hTFtarget","KnockTF","FIMO_JASPAR","PWMEnrich_JASPAR"), cor_DB = c("TCGA","GTEx"),tf = "STAT3")
results_inter <- intersections(results)
plot_venn(results_inter)
}
}
