% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{refgene}
\alias{refgene}
\docType{data}
\title{RefSeq transcript to gene symbol map}
\usage{data(refgene)}
\format{A data.frame with 53318 rows and 2 columns (Target, symbol).}
\description{
RefSeq transcript accessions mapped to gene symbols.
}
\source{RefSeq.}
\keyword{datasets}
