% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tcga_gtex}
\alias{tcga_gtex}
\docType{data}
\title{Sample metadata used for pan-tissue correlation analysis}
\usage{data(tcga_gtex)}
\format{A data.frame with 18102 rows and 5 columns (sample, tissue, type, type2, type1).}
\description{
Sample-level annotation (sample id, tissue, sample type) mapping expression samples of TCGA and GTEx used by \code{pantissue_cor_analysis()}.
}
\source{TCGA and GTEx sample annotations.}
\keyword{datasets}
