% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TF_Target_batch.R
\name{TF_Target_batch}
\alias{TF_Target_batch}
\title{Predict the target genes of TFs}
\usage{
TF_Target_batch(
  datasets = c("FIMO_JASPAR"),
  tfs = c("STAT3", "AHR"),
  TCGA_tissue = "COAD",
  GTEx_tissue = "Colon",
  cor_DB = c("TCGA", "GTEx"),
  cor_cutoff = 0.3,
  FIMO.score = 10,
  PWMEnrich.p = 0.1,
  cut.log2FC = 1,
  down.only = T,
  app = F
)
}
\arguments{
\item{datasets}{Including the TF-target regulatory data of 7 public TF online tools and the TF-targets prediction results analyzed by using FIMO and PWMEnrich.}

\item{tfs}{Transcription factor names}

\item{TCGA_tissue}{Cancer type in TCGA database, you can use tissue_type("TCGA") to abtain the tissue types.}

\item{GTEx_tissue}{Cancer type in GTEx database, you can use tissue_type("GTEx") to abtain the tissue types.}

\item{cor_DB}{The database used for the correlation analyze between TF and targets. You can use 2 databases, viz. TCGA (33 cancer types) and GTEx (30 normal tissue types).}

\item{cor_cutoff}{Threshold of correlation coefficient for correlation analysis.}

\item{FIMO.score}{Threshold of the score of the prediction TF-target results by using FIMO algorithm.}

\item{PWMEnrich.p}{Threshold of the p value of the prediction TF-target results by using PWMEnrich algorithm (smaller is better), default 0.1.}

\item{cut.log2FC}{Threshold of log2FC for KnockTF dataset.}

\item{down.only}{Logic value. If true, only the downregulated genes in TF knockout/knockdown cells were returned in KnockTF dataset.}

\item{app}{Logic value. TRUE only used in the shiny app.}
}
\description{
Predict the target genes of Transcription Factors in multiple TF-target prediction databases and correlation analysis.
}
\examples{
\dontrun{
results <- TF_Target_batch(datasets=c("hTFtarget","KnockTF"), cor_DB = c("TCGA","GTEx"),tf = c("STAT3","FOXM1"))
}
}
