% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{tissue}
\alias{tissue}
\docType{data}
\title{Tissue type lists for TCGA and GTEx}
\usage{data(tissue)}
\format{A list of two character vectors named "TCGA" and "GTEx".}
\description{
Lists of available TCGA cancer types (33) and GTEx tissue types (30) used as choices for correlation analyses, see \code{tissue_type()}.
}
\source{TCGA and GTEx.}
\keyword{datasets}
