% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tissue_type.R
\name{tissue_type}
\alias{tissue_type}
\title{Tissue type information}
\usage{
tissue_type(db = "TCGA")
}
\arguments{
\item{db}{"TCGA" or "GTEx".}
}
\description{
Get the tissue types in "TCGA" or "GTEx" database
}
\examples{
\dontrun{
tissue_type("GTEx")
}
}
