% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz_cor_results.R
\name{viz_cor_results}
\alias{viz_cor_results}
\title{Visualization of pan-tissue correlation analysis}
\usage{
viz_cor_results(
  cor_results,
  values = c("green", "black", "red"),
  alpha = 0.6,
  size = 3,
  r.cut = 0.3,
  p.cut = 0.05,
  max.overlaps = 10
)
}
\arguments{
\item{cor_results}{Correlation analysis results obtained from pantissue_cor_analysis() function.}

\item{values}{Color of the points， default c("green","black","red").}

\item{alpha}{Transparency of points， default 0.6.}

\item{size}{Size of points， default 3}

\item{r.cut}{Threshold of correlation coefficient， default 0.6.}

\item{p.cut}{Threshold of P value， default 0.05.}

\item{max.overlaps}{Max overlaps of the point labels, default 10.}
}
\description{
Visualization of pan-tissue correlation analysis using ggplot2.
}
\examples{
\dontrun{
cor_results <- pantissue_cor_analysis(Gene1 = "FOXM1",Gene2 = "GAPDH")
viz_cor_results(cor_results,
                values = c("black","red"))
}
}
