% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowerplot.R
\name{flowerplot}
\alias{flowerplot}
\title{Plotting folower plot}
\usage{
flowerplot(
  flower_dat,
  angle = 90,
  a = 1,
  b = 2,
  r = 1,
  ellipse_col_pal = "Set1",
  circle_col = "white",
  label_text_cex = 1
)
}
\arguments{
\item{flower_dat}{List of data.}

\item{ellipse_col_pal}{Color palette in RColorBrewer, default "Set1".}

\item{label_text_cex}{Font size, default 1.}
}
\description{
Plotting the folower plot for visualizing the intersection of more than 5 datasets.
}
