% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get TF-target data}
\usage{
get_data(table = "ENCODE", searchType = "Target", gene = "GAPDH")
}
\arguments{
\item{table}{One of the dataset in "hTFtarget","KnockTF","FIMO_JASPAR","PWMEnrich_JASPAR","ENCODE","CHEA","TRRUST","GTRD","ChIP_Atlas". For correlation analysis, "cor_"+tissue type was the table name, i.e. cor_TCGA, cor_Lung.}

\item{searchType}{TF or Target.}

\item{gene}{Any TF names for searchType == TF, and any gene symbols for searchType == Target}
}
\description{
Get the TF-target targeting results by using the api. All results saved in MySQL database.
}
\examples{
\dontrun{
results <- get_data(table = "ENCODE", searchType = "Target", gene = "GAPDH")
}
}
