% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_analysis.R, R/pantissue_cor_analysis.R
\name{pantissue_cor_analysis}
\alias{pantissue_cor_analysis}
\title{Correlation analysis}
\usage{
pantissue_cor_analysis(
  Gene1 = "FOXM1",
  Gene2 = "GAPDH",
  data_source = "TCGA",
  type = c("normal", "tumor"),
  cor_method = "pearson"
)

pantissue_cor_analysis(
  Gene1 = "FOXM1",
  Gene2 = "GAPDH",
  data_source = "TCGA",
  type = c("normal", "tumor"),
  cor_method = "pearson"
)
}
\arguments{
\item{Gene1}{Gene1 name.}

\item{Gene2}{Gene2 name.}

\item{data_source}{Data source used for correlation analysis, e.g. "TCGA", "GTEx" and "CCLE".}

\item{type}{Sample types used for correlation analysis, e.g. "normal" or/and "tumor".}

\item{cor_method}{Method used for correlation analysis, e.g. "pearson", "spearman".}
}
\description{
Correlation analysis between TF and target gene in pan-tissue in "TCGA", "GTEx" or "CCLE" databases.

Correlation analysis between TF and target in pan-tissues in "TCGA", "GTEx" and "CCLE" databases.
}
\examples{
\dontrun{
results_inter <- pantissue_cor_analysis(Gene1 = "FOXM1",Gene2 = "GAPDH")
}
}
