% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_target.R
\name{predict_target}
\alias{predict_target}
\title{Predict the target genes of TF}
\usage{
predict_target(
  datasets = c("hTFtarget", "KnockTF", "FIMO_JASPAR", "PWMEnrich_JASPAR", "ENCODE",
    "CHEA", "TRRUST", "GTRD", "ChIP_Atlas"),
  tf = "STAT3",
  TCGA_tissue = "COAD",
  GTEx_tissue = "Colon",
  cor_DB = c("TCGA", "GTEx"),
  cor_cutoff = 0.3,
  FIMO.score = 10,
  PWMEnrich.score = 10,
  cut.log2FC = 1,
  down.only = T,
  app = F
)
}
\arguments{
\item{datasets}{Including the TF-target regulatory data of 7 public TF online tools and the TF-targets prediction results analyzed by using FIMO and PWMEnrich.}

\item{tf}{Transcription factor name}

\item{TCGA_tissue}{Cancer type in TCGA database, you can use tissue_type("TCGA") to abtain the tissue types.}

\item{GTEx_tissue}{Cancer type in GTEx database, you can use tissue_type("GTEx") to abtain the tissue types.}

\item{cor_DB}{The database used for the correlation analyze between TF and targets. You can use 2 databases, viz. TCGA (33 cancer types) and GTEx (31 normal tissue types).}

\item{cor_cutoff}{Threshold of correlation coefficient for correlation analysis.}

\item{FIMO.score}{Threshold of the score of the prediction TF-target results by using FIMO algorithm.}

\item{PWMEnrich.score}{Threshold of the score of the prediction TF-target results by using PWMEnrich algorithm..}

\item{cut.log2FC}{Threshold of log2FC for KnockTF dataset.}

\item{down.only}{Logic value. If true, only the downregulated genes in TF knockout/knockdown cells were returned in KnockTF dataset.}

\item{app}{Logic value. TRUE only used in the shiny app.}
}
\description{
destination description  Predict the target genes of Transcription Factor in multiple TF-target prediction databases and correlation analysis.
}
\examples{
\dontrun{
results <- predict_target(datasets=c("hTFtarget","KnockTF","FIMO_JASPAR","PWMEnrich_JASPAR"), cor_DB = c("TCGA","GTEx"),tf = "STAT3")
}
}
